+++ lib/keycloak-admin/representation/camel_json.rb
@@ -5,7 +5,7 @@ module KeycloakAdmin
       if first_letter_in_uppercase
         lower_case_and_underscored_word.to_s.gsub(/\/(.?)/) { "::" + $1.upcase }.gsub(/(^|_)(.)/) { $2.upcase }
       else
-        lower_case_and_underscored_word.first + camelize(lower_case_and_underscored_word)[1..-1]
+        lower_case_and_underscored_word[0] + camelize(lower_case_and_underscored_word)[1..-1]
       end
     end
   end