+++ lib/keycloak-admin/representation/representation.rb
@@ -4,6 +4,10 @@ require_relative "camel_json"
 class Representation
   include ::KeycloakAdmin::CamelJson
 
+  def as_json
+    Hash[instance_variables.map { |ivar| [ivar.to_s[1..-1], instance_variable_get(ivar)] }]
+  end
+
   def to_json(options=nil)
     snaked_hash = as_json
     snaked_hash.keys.reduce({}) do |camelized_hash, key|