+++ lib/keycloak-admin/representation/role_representation.rb
@@ -1,0 +1,17 @@
+module KeycloakAdmin
+  class RoleRepresentation < Representation
+    attr_accessor :id,
+      :name,
+      :composite,
+      :client_role
+
+    def self.from_hash(hash)
+      role             = new
+      role.id          = hash["id"]
+      role.name        = hash["name"]
+      role.composite   = hash["composite"]
+      role.client_role = hash["clientRole"]
+      role
+    end
+  end
+end