+++ spec/client/realm_client_spec.rb
@@ -44,4 +44,112 @@ RSpec.describe KeycloakAdmin::RealmClient do
       end
     end
   end
+
+  describe "#list" do
+    before(:each) do
+      @realm_client = KeycloakAdmin.realm('master')
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '[{"id":"test_realm","realm":"test_realm"}]'
+    end
+
+    it "lists realms" do
+      realms = @realm_client.list
+      expect(realms.length).to eq 1
+      expect(realms[0].realm).to eq "test_realm"
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms", rest_client_options).and_call_original
+
+      realms = @realm_client.list
+      expect(realms.length).to eq 1
+      expect(realms[0].realm).to eq "test_realm"
+    end
+  end
+
+  describe "#delete" do
+    let(:realm_name) { "valid-realm" }
+
+    before(:each) do
+      @realm_client = KeycloakAdmin.realm(realm_name)
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:delete)
+    end
+
+    it "delete realm" do
+      expect(@realm_client.delete).to be_truthy
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm", rest_client_options).and_call_original
+
+      expect(@realm_client.delete).to be_truthy
+    end
+  end
+
+  describe "#save" do
+    let(:realm_name) { "valid-realm" }
+    let(:realm) { KeycloakAdmin::RealmRepresentation.from_hash(
+      "id" => realm_name,
+      "realm" => realm_name
+    )}
+
+    before(:each) do
+      @realm_client = KeycloakAdmin.realm(nil)
+
+      stub_token_client
+
+      expect_any_instance_of(RestClient::Resource).to receive(:post).with(realm.to_json, anything)
+    end
+
+    it "saves a realm" do
+      @realm_client.save(realm)
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms", rest_client_options).and_call_original
+
+      @realm_client.save(realm)
+    end
+  end
+
+  describe "#update" do
+    let(:realm_name) { "valid-realm" }
+    let(:realm_json) { { smtpServer: { host: 'test_host' } } }
+
+    before(:each) do
+      @realm_client = KeycloakAdmin.realm(realm_name)
+
+      stub_token_client
+      expect_any_instance_of(RestClient::Resource).to receive(:put).with(realm_json.to_json, anything)
+    end
+
+    it "updates a realm" do
+      @realm_client.update(realm_json)
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm", rest_client_options).and_call_original
+
+      @realm_client.update(realm_json)
+    end
+  end
 end