+++ spec/client/user_client_spec.rb
@@ -97,15 +97,42 @@ RSpec.describe KeycloakAdmin::TokenClient do
     end
   end
 
+  describe "#save" do
+    let(:realm_name) { "valid-realm" }
+    let(:user) { KeycloakAdmin::UserRepresentation.from_hash(
+      "username" => "test_username",
+      "createdTimestamp" => Time.now.to_i,
+    )}
+
+    before(:each) do
+      @user_client = KeycloakAdmin.realm(realm_name).users
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:post)
+    end
+
+    it "saves a user" do
+      expect(@user_client.save(user)).to eq user
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm/users", rest_client_options).and_call_original
+
+      expect(@user_client.save(user)).to eq user
+    end
+  end
+
   describe "#get" do
     let(:realm_name) { "valid-realm" }
-    before(:each) do
-      @user_client = KeycloakAdmin.realm("a_realm").users
 
-      allow_any_instance_of(KeycloakAdmin::TokenClient).to receive(:get).and_return KeycloakAdmin::TokenRepresentation.new(
-        'test_access_token', 'token_type', 'expires_in', 'refresh_token',
-        'refresh_expires_in', 'id_token', 'not_before_policy', 'session_state'
-      )
+    before(:each) do
+      @user_client = KeycloakAdmin.realm(realm_name).users
+
+      stub_token_client
       allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '{"username":"test_username","createdTimestamp":1559347200}'
     end
 
@@ -119,27 +146,31 @@ RSpec.describe KeycloakAdmin::TokenClient do
       allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
 
       expect(RestClient::Resource).to receive(:new).with(
-        "http://auth.service.io/auth/admin/realms/a_realm/users/test_user_id", rest_client_options).and_call_original
+        "http://auth.service.io/auth/admin/realms/valid-realm/users/test_user_id", rest_client_options).and_call_original
 
       user = @user_client.get('test_user_id')
       expect(user.username).to eq 'test_username'
     end
   end
 
-  describe "#delete" do
+  describe "#search" do
     let(:realm_name) { "valid-realm" }
-    before(:each) do
-      @user_client = KeycloakAdmin.realm("a_realm").users
+    let(:user) { KeycloakAdmin::UserRepresentation.from_hash(
+      "username" => "test_username",
+      "createdTimestamp" => Time.now.to_i,
+    )}
 
-      allow_any_instance_of(KeycloakAdmin::TokenClient).to receive(:get).and_return KeycloakAdmin::TokenRepresentation.new(
-        'test_access_token', 'token_type', 'expires_in', 'refresh_token',
-        'refresh_expires_in', 'id_token', 'not_before_policy', 'session_state'
-      )
-      allow_any_instance_of(RestClient::Resource).to receive(:delete)
+    before(:each) do
+      @user_client = KeycloakAdmin.realm(realm_name).users
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '[{"username":"test_username","createdTimestamp":1559347200}]'
     end
 
-    it "parses the response" do
-      @user_client.delete('test_user_id')
+    it "finds a user" do
+      users = @user_client.search("test_username")
+      expect(users.length).to eq 1
+      expect(users[0].username).to eq "test_username"
     end
 
     it "passes rest client options" do
@@ -147,7 +178,67 @@ RSpec.describe KeycloakAdmin::TokenClient do
       allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
 
       expect(RestClient::Resource).to receive(:new).with(
-        "http://auth.service.io/auth/admin/realms/a_realm/users/test_user_id", rest_client_options).and_call_original
+        "http://auth.service.io/auth/admin/realms/valid-realm/users", rest_client_options).and_call_original
+
+      users = @user_client.search("test_username")
+      expect(users.length).to eq 1
+      expect(users[0].username).to eq "test_username"
+    end
+  end
+
+  describe "#list" do
+    let(:realm_name) { "valid-realm" }
+    let(:user) { KeycloakAdmin::UserRepresentation.from_hash(
+      "username" => "test_username",
+      "createdTimestamp" => Time.now.to_i,
+    )}
+
+    before(:each) do
+      @user_client = KeycloakAdmin.realm(realm_name).users
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '[{"username":"test_username","createdTimestamp":1559347200}]'
+    end
+
+    it "lists users" do
+      users = @user_client.list
+      expect(users.length).to eq 1
+      expect(users[0].username).to eq "test_username"
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm/users", rest_client_options).and_call_original
+
+      users = @user_client.list
+      expect(users.length).to eq 1
+      expect(users[0].username).to eq "test_username"
+    end
+  end
+
+  describe "#delete" do
+    let(:realm_name) { "valid-realm" }
+
+    before(:each) do
+      @user_client = KeycloakAdmin.realm(realm_name).users
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:delete)
+    end
+
+    it "does not fail" do
+      expect(@user_client.delete('test_user_id')).to be_truthy
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm/users/test_user_id", rest_client_options).and_call_original
 
       @user_client.delete('test_user_id')
     end