+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.x.y] - Next version
+
+* List users
+* List clients
+* List groups, create/save a group
+* List roles, save a role
+* List realms, save/update/delete a realm
+* Get list of client role mappings for a user
+* Support passing rest client options for user save and search
+* Support using gem without ActiveSupport
+
 ## [0.7.0] - 2019-06-06
 
 Thanks to @vlad-ro